Solve year 2022 day 16
Part 2 took a while because it turns out my implementation of the Floyd-Warshall algorithm was wrong. I didn't realize the ordering of the loops was significant (and also, I didn't understand how the algorithm worked in light of ordering requirements -- now I do!). But the algorithm was correct enough to pass part 1 and the test case for part 2.

I only figured it out by trying a different permutation of the input file (sorting it ascending), which resulted in different solution for part 2, which should clearly be impossible. I combed through my code looking for something non-deterministic, but I eventually found that the distance matrixes as produced by `floyd_warshall` were different.

The type-hints in this solution use the typing imports (e.g. `List`) instead of the types (e.g. `list`) directly, because I was using some version of Pypy, which didn't support indexing the types directly.