Solve year 2022 day 20
This took ages because I didn't realize that the input numbers weren't unique. There were several edge cases around wrapping that I was trying to work out, but it turns out that you don't need to worry about them for the actual input (as long as you know that the input numbers aren't unique).

The solution as-is is pretty slow (takes ~1min for part 2?), but I derived it directly from my previous, terribly-messy code, so it was easiest to implement from that point.